Implement user-based permissions

+++ src/components/pages/StudentsPage.tsx
@@ -1,12 +1,18 @@
 // Copyright 2022 Social Fabric, LLC
 
-import React, { useContext, useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
+import {
+  createPermission,
+  deletePermission,
+  updatePermission,
+} from '../../auth/permissions';
 import {
   Appointment,
   AppointmentsContext,
   emptyAppointment,
 } from '../../data/appointments';
 import { emptyStudent, Student, StudentsContext } from '../../data/students';
+import { LoggedInUserContext } from '../../data/users';
 import CreateAppointmentModal from '../modals/CreateAppointmentModal';
 import CreateStudentModal from '../modals/CreateStudentModal';
 import EditStudentModal from '../modals/EditStudentModal';
@@ -20,6 +26,8 @@ const StudentsPage = () => {
     update: updateStudent,
   } = useContext(StudentsContext);
   const { add: addAppointment } = useContext(AppointmentsContext);
+  const { loggedInUser } = useContext(LoggedInUserContext);
+
   const [isCreateStudentModalOpen, setIsCreateStudentModalOpen] =
     useState<boolean>(false);
   const [isEditStudentModalOpen, setIsEditStudentModalOpen] =
@@ -29,36 +37,63 @@ const StudentsPage = () => {
     useState<boolean>(false);
   const [modalAppointment, setModalAppointment] =
     useState<Appointment>(emptyAppointment);
+  const [isCreateAppointmentAllowed, setIsCreateAppointmentAllowed] =
+    useState<boolean>(false);
+  const [isCreateStudentAllowed, setIsCreateStudentAllowed] =
+    useState<boolean>(false);
+  const [isDeleteStudentAllowed, setIsDeleteStudentAllowed] =
+    useState<boolean>(false);
+  const [isUpdateStudentAllowed, setIsUpdateStudentAllowed] =
+    useState<boolean>(false);
+
+  useEffect(() => {
+    setIsCreateAppointmentAllowed(
+      createPermission(loggedInUser.role, 'appointment')
+    );
+    setIsCreateStudentAllowed(createPermission(loggedInUser.role, 'student'));
+    setIsDeleteStudentAllowed(deletePermission(loggedInUser.role, 'student'));
+    setIsUpdateStudentAllowed(updatePermission(loggedInUser.role, 'student'));
+  }, [loggedInUser.role]);
 
   const handleStudentAdded = (student: Student) => {
-    addStudent(student);
+    if (isCreateStudentAllowed) {
+      addStudent(student);
+    }
   };
 
   const handleStudentEdited = (student: Student) => {
-    updateStudent(student);
+    if (isUpdateStudentAllowed) {
+      updateStudent(student);
+    }
   };
 
   const handleAppointmentAdded = (appointment: Appointment) => {
-    addAppointment(appointment);
+    if (isCreateAppointmentAllowed) {
+      addAppointment(appointment);
+    }
   };
 
   const onDeleteStudentClicked = (studentToDelete: Student) => {
-    if (window.confirm('Delete this student?')) {
+    if (isDeleteStudentAllowed && window.confirm('Delete this student?')) {
       deleteStudent(studentToDelete);
     }
   };
 
   const onEditStudentClicked = (studentToEdit: Student) => {
-    setModalStudent(studentToEdit);
-    setIsEditStudentModalOpen(true);
+    if (isUpdateStudentAllowed) {
+      setModalStudent(studentToEdit);
+      setIsEditStudentModalOpen(true);
+    }
   };
 
   const onAppointmentStudentClicked = (studentToSchedule: Student) => {
-    modalAppointment.participants = [studentToSchedule];
-    modalAppointment.counselorId =
-      studentToSchedule.studentRef.assignedCounselorId;
-    setModalAppointment(modalAppointment);
-    setIsCreateAppointmentModalOpen(true);
+    if (isCreateAppointmentAllowed) {
+      modalAppointment.participants = [studentToSchedule];
+      modalAppointment.counselorId =
+        studentToSchedule.studentRef.assignedCounselorId;
+      setModalAppointment(modalAppointment);
+      setIsCreateAppointmentModalOpen(true);
+    }
   };
 
   return (
@@ -68,26 +103,37 @@ const StudentsPage = () => {
       </nav>
       <h1>Students</h1>
       <>
-        <button type="button" onClick={() => setIsCreateStudentModalOpen(true)}>
-          Add Student
-        </button>
-        <CreateAppointmentModal
-          isOpen={isCreateAppointmentModalOpen}
-          onAppointmentAdded={handleAppointmentAdded}
-          onClose={() => setIsCreateAppointmentModalOpen(false)}
-          initialAppointment={modalAppointment}
-        />
-        <CreateStudentModal
-          isOpen={isCreateStudentModalOpen}
-          onStudentAdded={handleStudentAdded}
-          onClose={() => setIsCreateStudentModalOpen(false)}
-        />
-        <EditStudentModal
-          isOpen={isEditStudentModalOpen}
-          onStudentEdited={handleStudentEdited}
-          onClose={() => setIsEditStudentModalOpen(false)}
-          initialStudent={modalStudent}
-        />
+        {isCreateStudentAllowed && (
+          <>
+            <button
+              type="button"
+              onClick={() => setIsCreateStudentModalOpen(true)}
+            >
+              Add Student
+            </button>
+            <CreateStudentModal
+              isOpen={isCreateStudentModalOpen}
+              onStudentAdded={handleStudentAdded}
+              onClose={() => setIsCreateStudentModalOpen(false)}
+            />
+          </>
+        )}
+        {isCreateAppointmentAllowed && (
+          <CreateAppointmentModal
+            isOpen={isCreateAppointmentModalOpen}
+            onAppointmentAdded={handleAppointmentAdded}
+            onClose={() => setIsCreateAppointmentModalOpen(false)}
+            initialAppointment={modalAppointment}
+          />
+        )}
+        {isUpdateStudentAllowed && (
+          <EditStudentModal
+            isOpen={isEditStudentModalOpen}
+            onStudentEdited={handleStudentEdited}
+            onClose={() => setIsEditStudentModalOpen(false)}
+            initialStudent={modalStudent}
+          />
+        )}
         <StudentsTable
           onDeleteClicked={onDeleteStudentClicked}
           onEditClicked={onEditStudentClicked}
